pyproject.toml
[tool.poetry]
name = "django-labs-toolkit"
version = "0.1.0"
description = "A collection of Django apps created and used by Penn Labs"
authors = ["Penn Labs <contact@pennlabs.org>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pennlabs/django-labs-toolkit"
repository = "https://github.com/pennlabs/django-labs-toolkit"
documentation = "https://django-labs-toolkit.readthedocs.io/en/latest/"
classifiers = [
    "Framework :: Django",
]
include = ["CHANGELOG.md", "LICENSE"]
exclude = ["tox.ini"]
packages = [
    { include = "pennlabs" },
]

[tool.poetry.dependencies]
python = "^3.6"
Django = ">=2.2"
djangorestframework = "^3.13.1"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
flake8-isort = "^4.0.0"
flake8-quotes = "^3.2.0"
flake8-absolute-import = "^1.0"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
